Return a value if the browser is considered outdated or not
The outdatedBrowserRework function now returns a promise which allows the user to react to the current state (is browser outdated or is browser allowed). The callback passed to the resolve method of a promise gets a boolean value passed. True means that the browser is not outdated, false means that the browser is outdated.