Use custom object deep extension function.
node-extend-deep is not compatible with older browsers as it uses ES5
stuff. Also it includes support Node Buffers, for which we have no
need and that support pulls in 10KB of unneeded Buffer polyfill
bloat. The function has also been stripped to bare minimums, only
supporting extending objects with simple primitive values.